+++ src/constants/General.ts
@@ -1,0 +1,3 @@
+export const General = {
+  demoFolder: ".demo",
+};